@@ -2,7 +2,7 @@
 # Module
 #     setup.cfg
 # Copyright
-#     Copyright (C) 2020 Vladimir Roncevic <elektron.ronca@gmail.com>
+#     Copyright (C) 2020 - 2024 Vladimir Roncevic <elektron.ronca@gmail.com>
 #     daemonpy is free software: you can redistribute it and/or modify it
 #     under the terms of the GNU General Public License as published by the
 #     Free Software Foundation, either version 3 of the License, or
@@ -14,33 +14,24 @@
 #     You should have received a copy of the GNU General Public License along
 #     with this program. If not, see <http://www.gnu.org/licenses/>.
 # Info
-#     Defined setup for daemonpy package.
+#     Defines setup for daemonpy package.
 #
 
 [metadata]
 name = daemonpy
-version = 1.9.3
+version = 1.8.2
 author = Vladimir Roncevic
 author_email = elektron.ronca@gmail.com
 description = Creating Daemon process
-license=2020 Free software to use and distributed it.
+license=GPL 2024 Free software to use and distributed it.
 long_description = file: README.md
 long_description_content_type = text/markdown
 keywords=Unix, Linux, Daemon, Process
 platforms = any
 url = https://vroncevic.github.io/daemonpy
 classifiers =
-    Programming Language :: Python :: 2.7
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.1
-    Programming Language :: Python :: 3.2
-    Programming Language :: Python :: 3.3
-    Programming Language :: Python :: 3.4
-    Programming Language :: Python :: 3.5
-    Programming Language :: Python :: 3.6
-    Programming Language :: Python :: 3.7
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
     License :: OSI Approved :: GNU Lesser General Public License v2 (LGPLv2)
     License :: OSI Approved :: GNU Lesser General Public License v2 or later (LGPLv2+)
     License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)
@@ -50,6 +41,6 @@ classifiers =
 
 [options]
 packages = find:
-python_requires = >=2.7
+python_requires = >=3.10
 install_requires=
     ats-utilities
