@@ -1,42 +1,43 @@
 # -*- coding: UTF-8 -*-
 
 '''
- Module
-     daemon_usage.py
- Copyright
-     Copyright (C) 2020 Vladimir Roncevic <elektron.ronca@gmail.com>
-     daemonpy is free software: you can redistribute it and/or modify it
-     under the terms of the GNU General Public License as published by the
-     Free Software Foundation, either version 3 of the License, or
-     (at your option) any later version.
-     daemonpy is distributed in the hope that it will be useful, but
-     WITHOUT ANY WARRANTY; without even the implied warranty of
-     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-     See the GNU General Public License for more details.
-     You should have received a copy of the GNU General Public License along
-     with this program. If not, see <http://www.gnu.org/licenses/>.
- Info
-     Defined class DaemonUsage with attribute(s) and method(s).
-     Created API for daemon usage.
+Module
+    daemon_usage.py
+Copyright
+    Copyright (C) 2020 - 2024 Vladimir Roncevic <elektron.ronca@gmail.com>
+    daemonpy is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+    daemonpy is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+    See the GNU General Public License for more details.
+    You should have received a copy of the GNU General Public License along
+    with this program. If not, see <http://www.gnu.org/licenses/>.
+Info
+    Defines class DaemonUsage with attribute(s) and method(s).
+    Creates an API for daemon usage.
 '''
 
 import sys
+from typing import List
 
 try:
     from ats_utilities.checker import ATSChecker
     from ats_utilities.console_io.error import error_message
     from ats_utilities.console_io.verbose import verbose_message
     from ats_utilities.exceptions.ats_type_error import ATSTypeError
-    from ats_utilities.exceptions.ats_bad_call_error import ATSBadCallError
+    from ats_utilities.exceptions.ats_value_error import ATSValueError
 except ImportError as ats_error_message:
-    MESSAGE = '\n{0}\n{1}\n'.format(__file__, ats_error_message)
-    sys.exit(MESSAGE)  # Force close python ATS ##############################
+    # Force close python ATS ##################################################
+    sys.exit(f'\n{__file__}\n{ats_error_message}\n')
 
 __author__ = 'Vladimir Roncevic'
-__copyright__ = 'Copyright 2020, https://vroncevic.github.io/daemonpy'
-__credits__ = ['Vladimir Roncevic']
+__copyright__ = '(C) 2024, https://vroncevic.github.io/daemonpy'
+__credits__: List[str] = ['Vladimir Roncevic', 'Python Software Foundation']
 __license__ = 'https://github.com/vroncevic/daemonpy/blob/dev/LICENSE'
-__version__ = '1.9.3'
+__version__ = '1.8.2'
 __maintainer__ = 'Vladimir Roncevic'
 __email__ = 'elektron.ronca@gmail.com'
 __status__ = 'Updated'
@@ -44,99 +45,88 @@ __status__ = 'Updated'
 
 class DaemonUsage:
     '''
-        Defined class DaemonUsage with attribute(s) and method(s).
-        Created API for daemon usage.
+        Defines class DaemonUsage with attribute(s) and method(s).
+        Creates an API for daemon usage.
+
         It defines:
 
             :attributes:
-                | PKG_VERBOSE - console text indicator for process-phase.
-                | DAEMON_OPERATIONS - list of supported operations.
-                | __usage_status - daemon usage status.
+                | _PKG_VERBOSE - Console text indicator for process-phase.
+                | DAEMON_OPERATIONS - List of supported operations.
+                | _usage_status - Daemon usage status.
             :methods:
-                | __init__ - initial constructor.
-                | usage_status - property methods for set/get operations.
-                | check - checking usage of daemon process.
-                | __str__ - dunder method for object DaemonUsage.
+                | __init__ - Initials DaemonUsage constructor.
+                | usage_status - Property methods for set/get operations.
+                | check - Checks usage for daemon process.
     '''
 
-    PKG_VERBOSE = 'DAEMONPY::DAEMON_USAGE'
-    DAEMON_OPERATIONS = ['start', 'stop', 'restart']
+    _PKG_VERBOSE: str = 'DAEMONPY::DAEMON_USAGE'
+    DAEMON_OPERATIONS: List[str] = ['start', 'stop', 'restart']
 
-    def __init__(self, verbose=False):
+    def __init__(self, verbose: bool = False) -> None:
         '''
-            Initial constructor.
+            Initials DaemonUsage constructor.
 
-            :param verbose: enable/disable verbose option.
+            :param verbose: Enable/Disable verbose option
             :type verbose: <bool>
             :exceptions: None
         '''
-        verbose_message(DaemonUsage.PKG_VERBOSE, verbose, 'init usage')
-        self.__usage_status = 0
+        verbose_message(verbose, [f'{self._PKG_VERBOSE} init usage'])
+        self._usage_status: int = 0
 
     @property
-    def usage_status(self):
+    def usage_status(self) -> int:
         '''
             Property method for getting daemon usage status.
 
-            :return: daemon usage status.
+            :return: Daemon usage status
             :rtype: <int>
             :exceptions: None
         '''
-        return self.__usage_status
+        return self._usage_status
 
     @usage_status.setter
-    def usage_status(self, usage_status):
+    def usage_status(self, usage_status: int) -> None:
         '''
             Property method for setting daemon usage status.
 
-            :param usage_status: daemon usage status.
+            :param usage_status: Daemon usage status
             :type usage_status: <int>
             :exceptions: None
         '''
-        self.__usage_status = usage_status
+        self._usage_status = usage_status
 
-    def check(self, daemon_operation, verbose=False):
+    def check(self, daemon_operation: str, verbose: bool = False):
         '''
-            Checking usage of Daemon process.
+            Checks usage of Daemon process.
 
-            :param daemon_operation: daemon operation.
+            :param daemon_operation: Daemon operation
             :type daemon_operation: <str>
-            :param verbose: enable/disable verbose option.
+            :param verbose: Enable/Disable verbose option
             :type verbose: <bool>
-            :exceptions: ATSTypeError | ATSBadCallError
+            :exceptions: ATSTypeError | ATSValueError
         '''
-        checker, error, status = ATSChecker(), None, False
-        error, status = checker.check_params([
+        error_msg: str | None = None
+        error_id: int | None = None
+        checker: ATSChecker = ATSChecker()
+        error_msg, error_id = checker.check_params([
             ('str:daemon_operation', daemon_operation)
         ])
-        if status == ATSChecker.TYPE_ERROR:
-            raise ATSTypeError(error)
-        if status == ATSChecker.VALUE_ERROR:
-            raise ATSBadCallError(error)
-        verbose_message(DaemonUsage.PKG_VERBOSE, verbose, 'checking usage')
-        for index, option in enumerate(DaemonUsage.DAEMON_OPERATIONS):
+        if error_id == checker.TYPE_ERROR:
+            raise ATSTypeError(error_msg)
+        if not bool(daemon_operation):
+            raise ATSValueError('missing daemon operation')
+        verbose_message(verbose, [f'{self._PKG_VERBOSE} checking usage'])
+        for index, option in enumerate(self.DAEMON_OPERATIONS):
             if option == daemon_operation:
-                self.__usage_status = index
-        if daemon_operation not in DaemonUsage.DAEMON_OPERATIONS:
-            self.__usage_status = 127
-            error_message(
-                DaemonUsage.PKG_VERBOSE, 'usage: {0}'.format(
-                    '|'.join(DaemonUsage.DAEMON_OPERATIONS)
-                )
-            )
+                self._usage_status = index
+        if daemon_operation not in self.DAEMON_OPERATIONS:
+            self._usage_status = 127
+            error_message([
+                f'{self._PKG_VERBOSE} usage: {0}',
+                '|'.join(self.DAEMON_OPERATIONS)
+            ])
         verbose_message(
-            DaemonUsage.PKG_VERBOSE, verbose,
-            'usage status', self.__usage_status
-        )
-
-    def __str__(self):
-        '''
-            Dunder method for DaemonUsage.
-
-            :return: object in a human-readable format.
-            :rtype: <str>
-            :exceptions: None
-        '''
-        return '{0} ({1})'.format(
-            self.__class__.__name__, str(self.__usage_status)
+            verbose,
+            [f'{self._PKG_VERBOSE} usage status', self._usage_status]
         )
