@@ -1,26 +1,27 @@
 # -*- coding: UTF-8 -*-
 
 '''
- Module
-     unix_operations.py
- Copyright
-     Copyright (C) 2020 Vladimir Roncevic <elektron.ronca@gmail.com>
-     daemonpy is free software: you can redistribute it and/or modify it
-     under the terms of the GNU General Public License as published by the
-     Free Software Foundation, either version 3 of the License, or
-     (at your option) any later version.
-     daemonpy is distributed in the hope that it will be useful, but
-     WITHOUT ANY WARRANTY; without even the implied warranty of
-     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-     See the GNU General Public License for more details.
-     You should have received a copy of the GNU General Public License along
-     with this program. If not, see <http://www.gnu.org/licenses/>.
- Info
-     Defined class UnixOperations with attribute(s) and method(s).
-     Created API for operating Unix Like OS processes.
+Module
+    unix_operations.py
+Copyright
+    Copyright (C) 2020 - 2024 Vladimir Roncevic <elektron.ronca@gmail.com>
+    daemonpy is free software: you can redistribute it and/or modify it
+    under the terms of the GNU General Public License as published by the
+    Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+    daemonpy is distributed in the hope that it will be useful, but
+    WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+    See the GNU General Public License for more details.
+    You should have received a copy of the GNU General Public License along
+    with this program. If not, see <http://www.gnu.org/licenses/>.
+Info
+    Defines class UnixOperations with attribute(s) and method(s).
+    Creates an API for operating Unix Like OS processes.
 '''
 
 import sys
+from typing import List
 from os import fork, kill, remove
 from os.path import exists
 from signal import SIGTERM
@@ -31,16 +32,16 @@ try:
     from ats_utilities.console_io.error import error_message
     from ats_utilities.console_io.verbose import verbose_message
     from ats_utilities.exceptions.ats_type_error import ATSTypeError
-    from ats_utilities.exceptions.ats_bad_call_error import ATSBadCallError
+    from ats_utilities.exceptions.ats_value_error import ATSValueError
 except ImportError as ats_error_message:
-    MESSAGE = '\n{0}\n{1}\n'.format(__file__, ats_error_message)
-    sys.exit(MESSAGE)  # Force close python ATS ##############################
+    # Force close python ATS ##################################################
+    sys.exit(f'\n{__file__}\n{ats_error_message}\n')
 
 __author__ = 'Vladimir Roncevic'
-__copyright__ = 'Copyright 2020, https://vroncevic.github.io/daemonpy'
-__credits__ = ['Vladimir Roncevic']
+__copyright__ = '(C) 2024, https://vroncevic.github.io/daemonpy'
+__credits__: List[str] = ['Vladimir Roncevic', 'Python Software Foundation']
 __license__ = 'https://github.com/vroncevic/daemonpy/blob/dev/LICENSE'
-__version__ = '1.9.3'
+__version__ = '1.8.2'
 __maintainer__ = 'Vladimir Roncevic'
 __email__ = 'elektron.ronca@gmail.com'
 __status__ = 'Updated'
@@ -48,174 +49,147 @@ __status__ = 'Updated'
 
 class UnixOperations:
     '''
-        Defined class UnixOperations with attribute(s) and method(s).
-        Created API for operating Unix Like OS processes.
+        Defines class UnixOperations with attribute(s) and method(s).
+        Creates an API for operating Unix Like OS processes.
+
         It defines:
 
             :attributes:
-                | PKG_VERBOSE - console text indicator for process-phase.
-                | OS_TARGET - list of supported operating systems.
-                | NO_PROCESS - no such process message.
-                | __unix_status - unix status (True for unix like OS).
+                | _PKG_VERBOSE - Console text indicator for process-phase.
+                | _OS_TARGET - List of supported operating systems.
+                | _NO_PROCESS - No such process message.
+                | _unix_status - Unix status (True for unix like OS).
             :methods:
-                | __init__ - initial constructor.
-                | unix_status - property methods for set/get operations.
-                | first_fork - make sure that process is not group leader.
-                | second_fork - won't be started merely by opening a terminal.
-                | unix_kill - kill unix like OS process.
-                | __str__ - dunder method for object UnixOperations.
+                | __init__ - Initials UnixOperations constructor.
+                | unix_status - Property methods for set/get operations.
+                | first_fork - Makes sure that process is not group leader.
+                | second_fork - Won't be started merely by opening a terminal.
+                | unix_kill - Kills unix like OS process.
     '''
 
-    PKG_VERBOSE = 'DAEMONPY::UNIX_OPERATIONS'
-    OS_TARGET = ['linux', 'linux2']
-    NO_PROCESS = 'No such process'
+    _PKG_VERBOSE: str = 'DAEMONPY::UNIX_OPERATIONS'
+    _OS_TARGET: List[str] = ['linux', 'linux2']
+    _NO_PROCESS: str = 'No such process'
+    _SLEEP: float = 0.1
 
-    def __init__(self, verbose=False):
+    def __init__(self, verbose: bool = False) -> None:
         '''
-            Initial constructor.
+            Initials UnixOperations constructor.
 
+            :param verbose: Enable/Disable verbose option
+            :type verbose: <bool>
             :exceptions: None
         '''
         verbose_message(
-            UnixOperations.PKG_VERBOSE, verbose, 'init daemon operations'
+            verbose, [f'{self._PKG_VERBOSE} init daemon operations']
         )
-        self.__unix_status = any([
-            sys.platform == target for target in UnixOperations.OS_TARGET
-        ])
+        self._unix_status: bool = any(
+            sys.platform == os_target for os_target in self._OS_TARGET
+        )
 
     @property
-    def unix_status(self):
+    def unix_status(self) -> bool:
         '''
             Property method for getting unix like OS status.
 
-            :return: unix like OS status.
+            :return: Unix like OS status
             :rtype: <bool>
             :exceptions: None
         '''
-        return self.__unix_status
+        return self._unix_status
 
     @unix_status.setter
-    def unix_status(self, unix_status):
+    def unix_status(self, unix_status: bool) -> None:
         '''
             Property method for setting unix like OS status.
 
-            :param unix_status: unix like OS status.
+            :param unix_status: Unix like OS status
             :type unix_status: <bool>
             :exceptions: None
         '''
-        self.__unix_status = unix_status
+        self._unix_status = unix_status
 
-    def first_fork(self, verbose=False):
+    def first_fork(self, verbose: bool = False) -> None:
         '''
-            Make sure that process is not group leader.
+            Makes sure that process is not group leader.
 
-            :param verbose: enable/disable verbose option.
+            :param verbose: Enable/Disable verbose option
             :type verbose: <bool>
-            :exit code: 0 (success) | 1 (failed)
+            :exit code: 0 (success fork) | 1 (failed)
             :exceptions: None
         '''
-        if self.__unix_status:
-            try:
-                process_id = fork()
-                if process_id > 0:
-                    verbose_message(
-                        UnixOperations.PKG_VERBOSE, verbose, 'first fork'
-                    )
-                    sys.exit(0)
-            except OSError as os_error:
-                error_message(
-                    'fork #1 failed: {0} {1}\n'.format(
-                        os_error.errno, os_error.strerror
-                    )
-                )
-                sys.exit(1)
+        if self._unix_status:
+            if fork() > 0:
+                verbose_message(verbose, f'{self._PKG_VERBOSE} first fork')
+                sys.exit(0)
 
-    def second_fork(self, verbose=False):
+    def second_fork(self, verbose: bool = False) -> None:
         '''
             Won't be started merely by opening a terminal.
 
-            :param verbose: enable/disable verbose option.
+            :param verbose: Enable/Disable verbose option
             :type verbose: <bool>
-            :exit code: 0 (success) | 1 (failed)
+            :exit code: 0 (success fork) | 1 (failed)
             :exceptions: None
         '''
-        if self.__unix_status:
-            try:
-                process_id = fork()
-                if process_id > 0:
-                    verbose_message(
-                        UnixOperations.PKG_VERBOSE, verbose, 'second fork'
-                    )
-                    sys.exit(0)
-            except OSError as os_error:
-                error_message(
-                    'fork #2 failed: {0} {1}\n'.format(
-                        os_error.errno, os_error.strerror
-                    )
-                )
-                sys.exit(1)
+        if self._unix_status:
+            if fork() > 0:
+                verbose_message(verbose, [f'{self._PKG_VERBOSE} second fork'])
+                sys.exit(0)
 
-    def unix_kill(self, process_id, pid_file_path, verbose=False):
+    def unix_kill(
+        self, pid: int, pid_path: str, verbose: bool = False
+    ) -> bool:
         '''
-            Kill Unix Like OS process.
+            Kills Unix Like OS process.
 
-            :param process_id: process ID.
-            :type process_id: <int>
-            :param pid_file_path: PID file path.
-            :type pid_file_path: <str>
-            :param verbose: enable/disable verbose option.
+            :param pid: Process ID
+            :type pid: <int>
+            :param pid_path: PID file path
+            :type pid_path: <str>
+            :param verbose: Enable/Disable verbose option
             :type verbose: <bool>
-            :return: boolean status, True | False.
+            :return: True (success operation) | False
             :rtype: <bool>
-            :exceptions: ATSTypeError | ATSBadCallError
+            :exceptions: ATSTypeError | ATSValueError
         '''
-        checker, error, status, sleep_time = ATSChecker(), None, False, 0.1
-        error, status = checker.check_params([
-            ('int:process_id', process_id),
-            ('str:pid_file_path', pid_file_path)
+        error_msg: str | None = None
+        error_id: int | None = None
+        checker: ATSChecker = ATSChecker()
+        error_msg, error_id = checker.check_params([
+            ('int:pid', pid), ('str:pid_path', pid_path)
         ])
-        if status == ATSChecker.TYPE_ERROR:
-            raise ATSTypeError(error)
-        if status == ATSChecker.VALUE_ERROR:
-            raise ATSBadCallError(error)
-        if self.__unix_status:
+        if error_id == checker.TYPE_ERROR:
+            raise ATSTypeError(error_msg)
+        if not bool(pid):
+            raise ATSValueError('missing PID')
+        if not bool(pid_path):
+            raise ATSValueError('missing PID path')
+        status: bool = False
+        if self._unix_status:
             try:
                 verbose_message(
-                    UnixOperations.PKG_VERBOSE, verbose,
-                    'kill process {0}'.format(process_id)
+                    verbose, [f'{self._PKG_VERBOSE} kill process {pid}']
                 )
-                while 1:
-                    kill(process_id, SIGTERM)
-                    sleep(sleep_time)
+                while True:
+                    kill(pid, SIGTERM)
+                    sleep(self._SLEEP)
                     status = True
             except OSError as os_error:
                 os_error = str(os_error)
-                if os_error.find(UnixOperations.NO_PROCESS) > 0:
-                    if exists(pid_file_path):
+                if os_error.find(self._NO_PROCESS) > 0:
+                    if exists(pid_path):
                         verbose_message(
-                            UnixOperations.PKG_VERBOSE, verbose,
-                            '{0} with PID: {1}, removing pid file {2}'.format(
-                                UnixOperations.NO_PROCESS,
-                                process_id, pid_file_path
-                            )
+                            verbose,
+                            [
+                                f'{self._PKG_VERBOSE}',
+                                f'{self._NO_PROCESS}',
+                                f'with PID: {pid},',
+                                f'removing pid file {pid_path}'
+                            ]
                         )
-                        remove(pid_file_path)
+                        remove(pid_path)
                         status = True
                 else:
-                    error_message(
-                        UnixOperations.PKG_VERBOSE, '{0}'.format(os_error)
-                    )
-                    status = False
+                    error_message([f'{self._PKG_VERBOSE} {os_error}'])
         return status
-
-    def __str__(self):
-        '''
-            Dunder method for UnixOperations.
-
-            :return: object in a human-readable format.
-            :rtype: <str>
-            :exceptions: None
-        '''
-        return '{0} ({1})'.format(
-            self.__class__.__name__, str(self.__unix_status)
-        )
